Add the ability to upload and preview images
Add the stampOptions & setStampOptions to state in app.js. Add a custom
slider to set the stamp size, and an input of type file to upload
images. Create an empty image tag (to preview images), and a function
getImage that actually uploads the image, and displays it in the
preview.